Fix 'branch' deadlock problem
The solution: Don't add duplicate threads to cp.
when adding two new threads (x, y)  to 'cp' for a branch instruction, for
each thread check first whether a thread pointing to the same instruction
already exists in cp. This is implemented using a lookup table 'cp_lookup'
of size 'compiled->size', which tracks whether or not an instruction has
been executed for the current input character.