Improve rxvm_fsearch() performance by buffering file reads
Instead of calling fgetc() for every character, maintain a buffer of X
characters, and call fread() to fill it. This way, when the VM calls
getchar_file() to get a new character, for (X - 1) calls out of X the
function will simply return a character from the buffer. File I/O
operations are only incurred by 1 of every X calls to getchar_file().

This has some improvements on rxvm_fsearch completion time (tests
performed with a 1GB base64-encoded plain text file, and using a value
of 256 for X):

- Time to search entire file for a simple expression (contains no
  meta-characters):

  Before: ~9 seconds
  Now:    ~5 seconds

- Time to search entire file for a regular expression (contains
  meta-characters):

  Before: ~25 seconds
  After:  ~22 seconds