Two significant VM throughput optimisations
1. Instead of exiting the VM and calling again for each failed match (when
   using rxvm_fsearch, for example, this will happen for every character
   until a match begins), simply use a goto and jump back to the start
   of vm_execute().

   In addition to avoiding the potential overhead of invoking vm_execute()
   following every non-matching character, this also allows the initial
   (costly) memset call at the beginning of vm_execute, which is unneeded in
   this case.

2. After consuming a character, the two thread tables and their
   respective lookup tables must be swapped. However, if only a single
   thread was created and it failed (again,a common case for rxvm_fsearch),
   then we can do a much simpler and quicker version of the swap.

The combined effect of these two changes appears to be a 100% increase
in throughput, at least for the following simple test case:

I generate a plain text file like so (usually test with multiple files,
ranging in size from 1MB to 5GB)-

    > cat /dev/urandom | base64 | head -c<file_size_in_bytes>

Then, I execute the rxvm_fsearch example app, searching the file for an
expression that is sufficiently complicated to ensure it wouldn't be in
the test file by chance. Before this commit, throughput on my machine
was ranging between ~50-60 million chars. per sec. Now, it's going at ~100-125
million chars. per sec.