rxvm_fsearch: use Boyer-Moore on full regexps when possible
Detect when an expression contains a fixed substring suitable for a
boyer-moore search. If it does, then we can use the fast boyer-moore algorithm
to search for that substring, and if we find it, back up to the last newline
character and run vm_execute for full regexp matching.

Before:
- Searching 1GB text file for "eriknyquist": ~0.5 seconds
- Searching 1GB text file for "erik+nyquist+": ~7.0 seconds

Now:
- Searching 1GB text file for "eriknyquist": ~0.5 seconds
- Searching 1GB text file for "erik+nyquist+": ~0.7 seconds